requirements.txt
Django==3.1
djangorestframework==3.11.1
requests==2.24.0
django-filter==2.3.0
